hours to decimal

+++ backend/StudyRoom.API/Models/UserStats.cs
@@ -9,9 +9,9 @@ public class UserStats
     [Key]
     public Guid UserId { get; set; }
 
-    public int TotalStudyMinutes { get; set; }
+    public decimal TotalStudyMinutes { get; set; }
     public int SessionsCompleted { get; set; }
     public int DailyStreak { get; set; }
-    public int WeeklyStudyMinutes { get; set; }
+    public decimal WeeklyStudyMinutes { get; set; }
     public DateTime UpdatedAt { get; set; } = DateTime.UtcNow;
 }
